Put together Upcoming/Past Events Activity
New layouts for fragments were created.  Activity is set to have lists
of events plugged in later.  The page is a slider for Upcoming and Past.
Has more of an Android look.  In later versions a possibility to make it
look more like the mock-ups.  Added comment into Login Activity to allow
for one-click login while testing, remove when going LIVE.